requirements.txt
Flask>=2.0.0
python-dateutil>=2.8.0
requests>=2.31.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
